backend/requirements.txt
Django==3.2.15
django-cors-headers==3.13.0
django-import-export==3.2.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.13.1
djoser==2.2.0
drf-extra-fields==3.4.0
flake8==6.0.0
importlib-metadata==4.12.0
itypes==1.2.0
MarkupSafe==2.1.1
Pillow==9.2.0
psycopg2-binary==2.9.3
pyflakes==3.0.0
python-dotenv==1.0.0
PyYAML==6.0
pytz==2022.2
requests==2.28.1
requests-oauthlib==1.3.1
social-auth-app-django==5.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
gunicorn==20.1.0